wip - refactor /generate-word
