+++ app/cv/generate/word.js
@@ -1,11 +1,13 @@
 import {
   Document,
-  Paragraph,
-  TextRun,
-  BorderStyle,
-  AlignmentType,
   Packer
 } from 'docx'
+import { styles } from '../word/styles.js'
+import { createHeader } from '../word/sections/header.js'
+import { createEducation } from '../word/sections/education.js'
+import { createExperience } from '../word/sections/experience.js'
+import { createProjects } from '../word/sections/projects.js'
+import { createInterests } from '../word/sections/interests.js'
 
 export const generateWordDoc = async (formData, res) => {
   if (!formData) {
@@ -13,11 +15,40 @@ export const generateWordDoc = async (formData, res) => {
   }
 
   try {
-    const doc = new Document({
-      sections: [],
+    const header = createHeader(formData)
+    const education = createEducation(formData)
+    const experience = createExperience(formData)
+    const projects = createProjects(formData)
+    const interests = createInterests(formData)
 
+    const sections = [
+      ...header,
+      ...education,
+      ...experience,
+      ...projects,
+      ...interests
+    ]
+
+    const doc = new Document({
+      sections: [
+        {
+          properties: {},
+          children: sections
+        }
+      ],
+      styles
+    })
+
+    Packer.toBuffer(doc).then(buffer => {
+      res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.wordprocessingml.document')
+      res.setHeader('Content-Disposition', 'attachment; filename=resume.docx')
+      res.send(buffer)
+    }).catch(error => {
+      console.error('Word document packing error:', error)
+      res.status(500).send('Failed to pack Word document')
     })
   } catch (error) {
-
+    console.error('Word document generation error:', error)
+    res.status(500).send('Failed to generate Word document')
   }
 }
